#  Improve memorize interface