--- a/packages/slate/src/models/mark.js
+++ b/packages/slate/src/models/mark.js
@@ -214,9 +214,7 @@ Mark.prototype[MODEL_TYPES.MARK] = true
  * Memoize read methods.
  */
 
-memoize(Mark.prototype, ['getComponent'], {
-  takesArguments: true,
-})
+memoize(Mark.prototype, ['getComponent'])
 
 /**
  * Export.