--- a/packages/slate/src/models/text.js
+++ b/packages/slate/src/models/text.js
@@ -511,19 +511,13 @@ memoize(Text.prototype, ['getMarks', 'getMarksAsArray'], {
   takesArguments: false,
 })
 
-memoize(
-  Text.prototype,
-  [
-    'getDecoratedCharacters',
-    'getDecorations',
-    'getLeaves',
-    'getMarksAtIndex',
-    'validate',
-  ],
-  {
-    takesArguments: true,
-  }
-)
+memoize(Text.prototype, [
+  'getDecoratedCharacters',
+  'getDecorations',
+  'getLeaves',
+  'getMarksAtIndex',
+  'validate',
+])
 
 /**
  * Export.