… packages/slate/src/models/node.js