1. [DONE] use `func.length` to replace `takesArgument` 
2.  [DONE] Add `post` interface to infer and set cache in other nodes