@zhujinxuan 2018-03-20
with both #1714 and this post interface, the benchmark is 
```
benchmarks
    html-serializer
      deserialize
        9.08 → 9.31 ops/sec
      serialize
        706.17 → 1081.84 ops/sec (53% faster)
    plain-serializer
      deserialize
        391.95 → 570.18 ops/sec (45% faster)
      serialize
        7887.13 → 10603.34 ops/sec (34% faster)
    rendering
      normal
        91.71 → 126.01 ops/sec (37% faster)
    changes
      delete-backward
        119634.72 → 140803.69 ops/sec
      delete-forward
        23448.94 → 36411.25 ops/sec (55% faster)
      insert-text-by-key-multiple
        95.94 → 133.02 ops/sec (39% faster)
      insert-text-by-key
        919.46 → 1245.49 ops/sec (35% faster)
      insert-text
        899.50 → 1156.37 ops/sec
      normalize
        1971.23 → 2239.59 ops/sec
      split-block
        28.18 → 31.09 ops/sec
    models
      from-json
        90.44 → 114.11 ops/sec
      get-blocks-at-range
        15827.09 → 18976.65 ops/sec
      get-blocks
        660651.56 → 576647.34 ops/sec
      get-characters-at-range
        9457.75 → 7250.64 ops/sec
      get-characters
        9768.90 → 9683.52 ops/sec
      get-inlines-at-range
        339931.46 → 280573.95 ops/sec
      get-inlines
        1558645.38 → 1283291.08 ops/sec
      get-leaves
        1541913.39 → 3466946.24 ops/sec (125% faster) 🙌
      get-marks-at-range
        1465.62 → 1631.63 ops/sec
      get-marks
        2280.98 → 1990.81 ops/sec
      get-path
        884555.59 → 702686.34 ops/sec
      get-texts-at-range
        565170.54 → 464544.12 ops/sec
      get-texts
        1030002.46 → 997138.77 ops/sec
      has-node-multiple
        120752.77 → 93097.92 ops/sec
      has-node
        923524.68 → 753612.76 ops/sec
      to-json
        3990.51 → 3697.66 ops/sec
      update-node
        20440.03 → 17384.67 ops/sec

✨  Done in 57.15s.
```