@zhujinxuan 2018-03-20
Reason of faster getLeaves:
1. each time of getLeaves(decoration=[]); the decoration is associated with a new empty array.  Then it would not take advantage of cache
2. solution. when use the default decorations, takeArguments is considered false by 
```
original.length === 0 && args.length === 0
```