@zhujinxuan 2018-03-21
Update only use last 3 keys, and benchmark as when all possible apps are closed:

only this PR:

this PR and #1714