@ianstormtaylor 2018-03-21
@zhujinxuan please don't combine separate features in pull requests like this. I want to merge the removal of `takesArguments` immediately, but I'm unsure about the other one. And since they're combined I now can't do anything. If you split these into two separate pull requests I can consider them. Thanks