@zhujinxuan 2018-03-21
@ianstormtaylor Fixed.  The second feature is removed. Thank you.