@ianstormtaylor 2018-03-23
Thanks @zhujinxuan, a very clean PR in the end with minimal changes but that got the solution implemented well!